# docs/openapi-skeleton.yaml
...
openapi: 3.0.3
info:
  title: Content Hunter API
  description: |
    API HTTP dell'app Next.js. Autenticazione tramite cookie di sessione NextAuth
    (`credentials: include`). Per admin globale usare header `x-company-id`.
  version: "1.0.0"
servers:
  - url: https://{host}
    variables:
      host:
        default: localhost:3000
        description: Host dell'installazione (es. contenthunter.abreve.it)
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: |
        Cookie impostato dopo login. Nome esatto può variare (es. __Secure-next-auth.session-token in produzione HTTPS).
    companyHeader:
      type: apiKey
      in: header
      name: x-company-id
      description: Obbligatorio per admin globale quando si opera su dati di un'azienda.
  parameters:
    CompanyIdHeader:
      name: x-company-id
      in: header
      required: false
      schema:
        type: integer
      description: Contesto azienda per utenti admin globali.
paths:
...
  /api/products:
    get:
      tags: [products]
      summary: Lista prodotti (filtri in query — verificare route.ts)
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/CompanyIdHeader"
      responses:
        "200":
          description: OK
    post:
      tags: [products]
      summary: Crea prodotto
      security:
        - sessionCookie: []
      responses:
        "201":
          description: Creato
...
security:
  - sessionCookie: []